+++ pkg/cookiebanner/tracker_mapping_agent_test.go
@@ -130,32 +130,6 @@ func TestNameIsCookieDatabaseAggregator(t *testing.T) {
 	}
 }
 
-func TestNormalizeAlnum(t *testing.T) {
-	t.Parallel()
-
-	tests := []struct {
-		name     string
-		input    string
-		expected string
-	}{
-		{name: "letters lowercased", input: "Letaido", expected: "letaido"},
-		{name: "strips punctuation and spaces", input: "letaido.com - Inc", expected: "letaidocominc"},
-		{name: "keeps digits", input: "auth0", expected: "auth0"},
-		{name: "empty", input: "", expected: ""},
-		{name: "only punctuation", input: "-_.:", expected: ""},
-	}
-
-	for _, tt := range tests {
-		t.Run(
-			tt.name,
-			func(t *testing.T) {
-				t.Parallel()
-				assert.Equal(t, tt.expected, normalizeAlnum(tt.input))
-			},
-		)
-	}
-}
-
 func TestBuildAgentPrompt(t *testing.T) {
 	t.Parallel()
